+++ src/WFC/WavePossition.cs
@@ -2,10 +2,10 @@
 
 public class WavePossition
 {
-    public int[] entropy { get; set; }
+    public byte[] entropy { get; set; }
     public bool collapsed => entropy.Length == 1;
 
-    public WavePossition(int[] entropy)
+    public WavePossition(byte[] entropy)
     {
         this.entropy = entropy;
     }
